# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AIAE Operational Hub API
  version: "1"
  description: >
    Backend API for AI Digital Operational Hub RBAC foundation. Authentication is Clerk SSO;
    the backend validates Clerk-issued JWTs against Clerk JWKS. OpenAPI is the source of truth
    for all exposed REST endpoints.
servers:
  - url: /
    description: Endpoint paths are fully qualified under /api/v1.
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/dictionary/rbac/roles:
    get:
      tags:
        - dictionary
      operationId: listRoles
      description: Lists active RBAC role dictionary rows. Requires manage-roles permission.
      responses:
        '200':
          description: Active role dictionary rows were found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while loading role dictionary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/dictionary/rbac/scope-types:
    get:
      tags:
        - dictionary
      operationId: listScopeTypes
      description: Lists active RBAC scope-type dictionary rows. Requires manage-roles permission.
      responses:
        '200':
          description: Active scope-type dictionary rows were found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScopeTypeV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while loading scope-type dictionary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/dictionary/statuses:
    get:
      tags:
        - dictionary
      operationId: listStatuses
      description: Lists the lifecycle status dictionary used by the user-management filters. Requires manage-roles permission.
      responses:
        '200':
          description: Status dictionary rows were returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while loading the status dictionary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/rbac/users/search:
    post:
      tags:
        - rbac
      operationId: searchUsers
      description: >
        Searches Hub users with their single active role, backing the user-management table.
        Sorting and filtering are applied server-side; results are paginated. Requires manage-roles permission.
      parameters:
        - $ref: '#/components/parameters/PageNumberParam'
        - $ref: '#/components/parameters/PageSizeParam'
      requestBody:
        description: Filter and sort criteria for the user search.
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HubUserSearchRequestV1'
      responses:
        '200':
          description: A page of Hub users was resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HubUserPageResponseV1'
        '400':
          description: The paging parameters or search criteria failed validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while searching users.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/rbac/users/{userId}/role-assignments:
    get:
      tags:
        - rbac
      operationId: listRoleAssignments
      description: Lists active role assignments for the specified Hub user. Requires manage-roles permission.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The hub_users.id of the user whose active assignments should be returned.
      responses:
        '200':
          description: Active role assignments were found for the specified user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleAssignmentV1'
        '400':
          description: The specified userId parameter has invalid type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while loading role assignments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
    post:
      tags:
        - rbac
      operationId: assignRole
      description: Assigns a role to the specified Hub user within the requested scope. Requires manage-roles permission.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The hub_users.id of the target user receiving the role assignment.
      requestBody:
        description: Role assignment request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequestV1'
      responses:
        '201':
          description: The role assignment was created, or an equivalent active assignment already existed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentV1'
        '400':
          description: Validation errors of incoming DTO or RBAC business-rule errors occurred.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
                  - $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '404':
          description: Target Hub user was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while assigning the role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
  /api/v1/rbac/users/{userId}/role-assignments/{assignmentId}:
    delete:
      tags:
        - rbac
      operationId: revokeRole
      description: Revokes an active role assignment for the specified Hub user. Requires manage-roles permission.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The hub_users.id of the target user whose assignment should be revoked.
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: The hub_role_assignments.id of the assignment to revoke.
      responses:
        '204':
          description: Assignment was revoked or was already absent.
        '400':
          description: The specified userId or assignmentId parameter has invalid type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubValidationExceptionResponseV1'
        '401':
          description: The authorization token not provided or expired.
        '403':
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
        '500':
          description: Exception occurred while revoking the role assignment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalHubApiExceptionResponseV1'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT bearer token validated against Clerk JWKS.
  parameters:
    PageNumberParam:
      name: pageNumber
      in: query
      required: false
      description: One-based page number to return.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    RoleV1:
      type: object
      description: Representation of an RBAC role dictionary row.
      properties:
        id:
          type: integer
          format: int64
          description: Internal hub_roles.id.
        role_code:
          type: string
          description: Stable role code used by backend authorization logic.
        display_name:
          type: string
          description: Human-readable role name.
        description:
          type: string
          description: Human-readable role description.
        future:
          type: boolean
          description: Indicates that the role is seeded for future use and is not yet generally assignable.
        status:
          type: string
          description: Current role dictionary status.
      required:
        - id
        - role_code
        - display_name
        - status
    ScopeTypeV1:
      type: object
      description: Representation of an RBAC scope-type dictionary row.
      properties:
        id:
          type: integer
          format: int64
          description: Internal hub_scope_types.id.
        scope_code:
          type: string
          description: Stable scope code defining how role assignment scope_id is interpreted.
        display_name:
          type: string
          description: Human-readable scope type name.
        description:
          type: string
          description: Human-readable scope type description.
        status:
          type: string
          description: Current scope-type dictionary status.
      required:
        - id
        - scope_code
        - display_name
        - status
    StatusV1:
      type: object
      description: A lifecycle status dictionary row.
      properties:
        code:
          type: string
          description: Stable status code stored on the entity (e.g. ACTIVE).
        display_name:
          type: string
          description: Human-readable status label.
      required:
        - code
        - display_name
    HubUserSummaryV1:
      type: object
      description: Hub user row for the user-management table, including the single active role.
      properties:
        hub_user_id:
          type: integer
          format: int64
          description: Internal hub_users.id.
        full_name:
          type: string
          description: User display name.
        email:
          type: string
          description: User email address.
        status:
          type: string
          description: User lifecycle status.
        role_code:
          type: string
          description: The user's current active role code, or null when no role is assigned.
        team_id:
          type: integer
          format: int64
          description: The team id scoped by the user's active role, or null when not TEAM-scoped.
      required:
        - hub_user_id
        - email
    RoleAssignmentV1:
      type: object
      description: Representation of a user's role assignment within a scope.
      properties:
        id:
          type: integer
          format: int64
          description: Internal hub_role_assignments.id.
        user_id:
          type: integer
          format: int64
          description: Internal hub_users.id receiving the role assignment.
        role_code:
          type: string
          description: Assigned role code.
        scope_code:
          type: string
          description: Scope code controlling how scope_id is interpreted.
        scope_id:
          type: integer
          format: int64
          description: Scoped entity id, or null for unscoped assignments such as ALL.
        status:
          type: string
          description: Current assignment status.
      required:
        - id
        - user_id
        - role_code
        - scope_code
        - status
    AssignRoleRequestV1:
      type: object
      description: Request body used to assign a role to a Hub user.
      properties:
        role_code:
          type: string
          description: Role code to assign.
        scope_code:
          type: string
          description: Scope code for the assignment.
        scope_id:
          type: integer
          format: int64
          description: Scoped entity id, or absent for unscoped assignments.
      required:
        - role_code
        - scope_code
    OperationalHubValidationExceptionResponseV1:
      type: object
      description: Representation of validation exception occurred in REST API format.
      properties:
        errors:
          description: The list of validation errors occurred.
          type: array
          items:
            $ref: '#/components/schemas/FieldToErrorResponseV1'
        timestamp:
          type: string
          format: date-time
          description: The timestamp exception occurred.
        correlationId:
          type: string
          description: The log stream UUID.
      required:
        - errors
        - timestamp
        - correlationId
    FieldToErrorResponseV1:
      type: object
      description: Representation of validation error bound to a field or semantic request part.
      properties:
        code:
          type: string
          description: The code of validation error occurred.
        field:
          type: string
          description: The field or semantic request part not passed validation.
        error:
          type: string
          description: The validation error occurred.
      required:
        - code
        - field
        - error
    OperationalHubApiExceptionResponseV1:
      type: object
      description: Representation of exception occurred in REST API format.
      properties:
        code:
          type: string
          description: The code of exception occurred.
        message:
          type: string
          description: The message of exception occurred.
        timestamp:
          type: string
          format: date-time
          description: The timestamp exception occurred.
        correlationId:
          type: string
          description: The log stream UUID.
      required:
        - code
        - message
        - timestamp
        - correlationId
    DirectionEnumV1:
      type: string
      description: Sort direction applied to a single sortable field.
      enum:
        - ASC
        - DESC
      default: ASC
    FilterOperationEnumV1:
      type: string
      description: >
        Operation used to match a filter value against a field. CONTAINS performs a substring
        match (used for free-text columns); EQUALS performs an exact match (used for code or
        enumerated columns).
      enum:
        - CONTAINS
        - EQUALS
    PageResponseV1:
      type: object
      description: Base pagination metadata shared by all paged responses.
      properties:
        pageNumber:
          type: integer
          format: int32
          description: One-based number of the returned page.
        pageSize:
          type: integer
          format: int32
          description: Maximum number of items per page.
        totalElements:
          type: integer
          format: int64
          description: Total number of items matching the query across all pages.
        totalPages:
          type: integer
          format: int32
          description: Total number of pages available for the query.
      required:
        - pageNumber
        - pageSize
        - totalElements
        - totalPages
    HubUserSortFieldEnumV1:
      type: string
      description: >
        Sortable fields of the user-management table. The active role code is filterable but not
        sortable, as it is resolved through a separate role-assignment lookup.
      enum:
        - HUB_USER_ID
        - FULL_NAME
        - EMAIL
        - STATUS
    HubUserFilterFieldEnumV1:
      type: string
      description: Filterable fields of the user-management table.
      enum:
        - HUB_USER_ID
        - FULL_NAME
        - EMAIL
        - STATUS
        - ROLE_CODE
    HubUserSortFieldV1:
      type: object
      description: Single-field sort directive for the user search.
      properties:
        field:
          $ref: '#/components/schemas/HubUserSortFieldEnumV1'
        direction:
          $ref: '#/components/schemas/DirectionEnumV1'
      required:
        - field
    HubUserFilterFieldV1:
      type: object
      description: Single-field filter directive for the user search.
      properties:
        field:
          $ref: '#/components/schemas/HubUserFilterFieldEnumV1'
        value:
          type: string
          description: The value matched against the field using the given operation.
        operation:
          $ref: '#/components/schemas/FilterOperationEnumV1'
        caseSensitive:
          type: boolean
          description: Whether the match is case-sensitive.
          default: false
      required:
        - field
        - value
        - operation
    HubUserSearchRequestV1:
      type: object
      description: Filter and sort criteria for searching Hub users.
      properties:
        filters:
          type: array
          description: Filters applied additively (AND) to the user search.
          items:
            $ref: '#/components/schemas/HubUserFilterFieldV1'
        sorting:
          $ref: '#/components/schemas/HubUserSortFieldV1'
    HubUserPageResponseV1:
      description: A page of Hub user rows for the user-management table.
      allOf:
        - $ref: '#/components/schemas/PageResponseV1'
        - type: object
          properties:
            content:
              type: array
              description: The Hub user rows on this page.
              items:
                $ref: '#/components/schemas/HubUserSummaryV1'
          required:
            - content
...